Styling cleanup with react-native-paper; make ObsEdit into section components
