+++ package.json
@@ -48,6 +48,7 @@
     "react-native-modal": "^13.0.1",
     "react-native-modal-datetime-picker": "^13.1.0",
     "react-native-network-logger": "^1.12.0",
+    "react-native-paper": "^4.12.1",
     "react-native-permissions": "^3.3.1",
     "react-native-picker-select": "^8.0.4",
     "react-native-reanimated": "^2.7.0",
@@ -56,6 +57,7 @@
     "react-native-screens": "^3.8.0",
     "react-native-sensitive-info": "^6.0.0-alpha.9",
     "react-native-uuid": "^2.0.1",
+    "react-native-vector-icons": "^9.1.0",
     "react-native-vision-camera": "^2.13.0",
     "react-spring": "^8.0.27",
     "react-tinder-card": "^1.4.5",
